fix(core/vmm): fixe use of fences on VMM install
First a fence was required to enforce the dependency on
`install_info_ready` being set calling vmm_vm_install as there is no
real data dependency.

Second, we move the write fence which ensures the write to the page
table is visible to within the install function. It is it responsibility
to ensure it returns with its effects visible.

Signed-off-by: Jose Martins <josemartins90@gmail.com>